move the elements associated with a user's registration to auth (services, controller, ..) and add auth route
