+++ src/auth/application/useCases/SignUp_useCase.ts
@@ -1,10 +1,10 @@
 import { ExistingEmailError } from "../../../shared/errors/ExistingEmailError";
-import { User } from "../../domain/entities/User";
-import { UserRepository } from "../../domain/repositories/IUserRepository";
+import { User } from "../../../users/domain/entities/User";
+import { UserRepository } from "../../../users/domain/repositories/IUserRepository";
 import { BcryptPasswordHasher } from "../services/Bcrypt_service";
 
+export class SignUpUseCase {
 
-export class RegisterUserUseCase {
   constructor(
     private readonly userRepository: UserRepository,
     private readonly bcryptService: BcryptPasswordHasher
