+++ src/auth/infraestructure/routes/Auth_routes.ts
@@ -1,0 +1,6 @@
+import express from 'express';
+import { signUpController } from '../dependencies';
+
+export const authRouter = express.Router();
+
+authRouter.post("/sign-up", signUpController.execute);
