+++ src/users/infraestructure/routes/User_routes.ts
@@ -1,6 +1,3 @@
 import express from 'express';
-import { registerUserController } from '../dependencies';
 
 export const userRouter = express.Router()
-
-userRouter.post("/", registerUserController.execute);
